feat: adapt SolverFactory implementations around
