+++ solve-classic/src/commonMain/kotlin/it/unibo/tuprolog/solve/classic/ClassicSolverFactory.kt
@@ -4,7 +4,9 @@ import it.unibo.tuprolog.solve.MutableSolver
 import it.unibo.tuprolog.solve.Solver
 import it.unibo.tuprolog.solve.SolverFactory
 import it.unibo.tuprolog.solve.channel.InputChannel
+import it.unibo.tuprolog.solve.channel.InputStore
 import it.unibo.tuprolog.solve.channel.OutputChannel
+import it.unibo.tuprolog.solve.channel.OutputStore
 import it.unibo.tuprolog.solve.classic.stdlib.DefaultBuiltins
 import it.unibo.tuprolog.solve.exception.Warning
 import it.unibo.tuprolog.solve.flags.FlagStore
@@ -16,6 +18,15 @@ object ClassicSolverFactory : SolverFactory {
 
     override val defaultBuiltins: Library
         get() = DefaultBuiltins
+
+    override fun solverOf(
+        libraries: Runtime,
+        flags: FlagStore,
+        staticKb: Theory,
+        dynamicKb: Theory,
+        inputs: InputStore,
+        outputs: OutputStore
+    ): Solver = ClassicSolver(libraries, flags, staticKb, dynamicKb, inputs, outputs)
 
     override fun solverOf(
         libraries: Runtime,
@@ -38,4 +49,13 @@ object ClassicSolverFactory : SolverFactory {
         stdErr: OutputChannel<String>,
         warnings: OutputChannel<Warning>
     ): MutableSolver = MutableClassicSolver(libraries, flags, staticKb, dynamicKb, stdIn, stdOut, stdErr, warnings)
+
+    override fun mutableSolverOf(
+        libraries: Runtime,
+        flags: FlagStore,
+        staticKb: Theory,
+        dynamicKb: Theory,
+        inputs: InputStore,
+        outputs: OutputStore
+    ): MutableSolver = MutableClassicSolver(libraries, flags, staticKb, dynamicKb, inputs, outputs)
 }
