+++ solve-concurrent/src/commonMain/kotlin/it/unibo/tuprolog/solve/concurrent/ConcurrentSolverFactory.kt
@@ -4,7 +4,9 @@ import it.unibo.tuprolog.solve.MutableSolver
 import it.unibo.tuprolog.solve.Solver
 import it.unibo.tuprolog.solve.SolverFactory
 import it.unibo.tuprolog.solve.channel.InputChannel
+import it.unibo.tuprolog.solve.channel.InputStore
 import it.unibo.tuprolog.solve.channel.OutputChannel
+import it.unibo.tuprolog.solve.channel.OutputStore
 import it.unibo.tuprolog.solve.concurrent.stdlib.DefaultBuiltins
 import it.unibo.tuprolog.solve.exception.Warning
 import it.unibo.tuprolog.solve.flags.FlagStore
@@ -15,6 +17,15 @@ import it.unibo.tuprolog.theory.Theory
 object ConcurrentSolverFactory : SolverFactory {
     override val defaultBuiltins: Library
         get() = DefaultBuiltins
+
+    override fun solverOf(
+        libraries: Runtime,
+        flags: FlagStore,
+        staticKb: Theory,
+        dynamicKb: Theory,
+        inputs: InputStore,
+        outputs: OutputStore
+    ): Solver = ConcurrentSolverImpl(libraries, flags, staticKb, dynamicKb, inputs, outputs)
 
     override fun solverOf(
         libraries: Runtime,
@@ -37,4 +48,13 @@ object ConcurrentSolverFactory : SolverFactory {
         stdErr: OutputChannel<String>,
         warnings: OutputChannel<Warning>
     ): MutableSolver = MutableConcurrentSolver(libraries, flags, staticKb, dynamicKb, stdIn, stdOut, stdErr, warnings)
+
+    override fun mutableSolverOf(
+        libraries: Runtime,
+        flags: FlagStore,
+        staticKb: Theory,
+        dynamicKb: Theory,
+        inputs: InputStore,
+        outputs: OutputStore
+    ): MutableSolver = MutableConcurrentSolver(libraries, flags, staticKb, dynamicKb, inputs, outputs)
 }
